errorfix: fix react/redux error on undefined passed to reducer | fix eslint error: no anonymous functions
exact error message:
Uncaught Error: The slice reducer for key "auth" returned undefined during initialization. If the state
passed to the reducer is undefined, you must explicitly return the initial state. The initial state may
not be undefined. If you don't want to set a value for this reducer, you can use null instead of
undefined.